requirements.txt
flask
flask-cors
python-dotenv
PyPDF2
sentence-transformers
chromadb
langchain
langchain-google-genai
google-genai
neo4j
spacy
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
